requirements.txt
requests~=2.32.4
gradio~=5.34.2
pandas~=2.3.0
chardet~=5.2.0
python-docx~=1.1.0
PyPDF2~=3.0.1
python-pptx~=0.6.23
